Move InspectSubtitle to grafana/ui
