+++ packages/grafana-ui/src/components/Inspector/types.ts
