+++ public/app/features/dashboard/components/Inspector/InspectContent.tsx
@@ -1,16 +1,14 @@
 import React, { useState } from 'react';
 import { DataSourceApi, PanelData, PanelPlugin } from '@grafana/data';
 import { getTemplateSrv } from '@grafana/runtime';
-import { CustomScrollbar, Drawer, TabContent } from '@grafana/ui';
+import { CustomScrollbar, Drawer, TabContent, InspectSubtitle, InspectTab } from '@grafana/ui';
 import { getPanelInspectorStyles } from './styles';
-import { InspectSubtitle } from './InspectSubtitle';
 import { InspectDataTab } from './InspectDataTab';
 import { InspectMetadataTab } from './InspectMetadataTab';
 import { InspectJSONTab } from './InspectJSONTab';
 import { InspectErrorTab } from './InspectErrorTab';
 import { InspectStatsTab } from './InspectStatsTab';
 import { QueryInspector } from './QueryInspector';
-import { InspectTab } from './types';
 import { DashboardModel, PanelModel } from '../../state';
 import { GetDataOptions } from '../../../query/state/PanelQueryRunner';
 
@@ -94,7 +92,9 @@ export const InspectContent: React.FC<Props> = ({
           )}
           {activeTab === InspectTab.Error && <InspectErrorTab error={error} />}
           {data && activeTab === InspectTab.Stats && <InspectStatsTab data={data} timeZone={dashboard.getTimezone()} />}
-          {data && activeTab === InspectTab.Query && <QueryInspector panel={panel} data={data.series} />}
+          {data && activeTab === InspectTab.Query && (
+            <QueryInspector panel={panel} data={data.series} onRefreshQuery={() => panel.refresh()} />
+          )}
         </TabContent>
       </CustomScrollbar>
     </Drawer>
